backend/requirements.txt
# Core Dependencies - Checked for v4.2 Modularity
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click>=8.1.0
cloudscraper==1.2.71
requests==2.32.5
curl_cffi>=0.12.0
python-docx==1.2.0
fpdf2>=2.8.0
openpyxl==3.1.5
python-dotenv==1.2.2
lxml==6.0.2
xlrd==2.0.1
cycler
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.135.1
filetype==1.2.0
fonttools==4.61.1
frozendict==2.4.7
google-auth==2.48.0
google-genai==1.66.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.17
langchain-google-genai==4.2.1
langsmith==0.7.12
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
# peewee==4.0.1 (Removed)
platformdirs==4.9.2
protobuf==7.34.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==0.52.1
tefas==1.0.1
tenacity==9.1.4
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
xxhash==3.6.0
yahooquery==2.4.1
zstandard==0.25.0
pytest==8.1.1
pytest-asyncio==0.23.6

# v6.3: GitHub Actions CI & Core Logic Verified

# v6.4: Optimizer & ML Hardening

# v7.3: Telegram Push & Portfolio Optimization

# v7.4: Risk Analyzer & Stress Testing

yfinance
pandas
numpy
PyJWT

# v7.5: Multi-Agent Architecture (Analyst/Researcher/CIO)
# v7.6: Advanced Risk Metrics (VaR & MaxDD)
# v7.7: Paper Trading Execution Logging Engine
# v7.8: Autonomous Background Alerting System
# v8.0: SOLID FastAPI Routers & ES6 Modular Frontend

